+++ .eslintrc.json
@@ -1,19 +1,19 @@
 {
     "env": {
-        "browser": true,
-        "es6": true
-    },
-    "extends": [
-        "airbnb-base"
-    ],
-    "globals": {
-        "Atomics": "readonly",
-        "SharedArrayBuffer": "readonly"
+      "browser": true,
+      "es6": true,
+      "jest": true
     },
     "parserOptions": {
-        "ecmaVersion": 2018,
-        "sourceType": "module"
+      "ecmaVersion": 2018,
+      "sourceType": "module"
     },
+    "extends": ["airbnb-base"],
     "rules": {
+      "linebreak-style": "off",
+      "no-shadow": "off",
+      "no-param-reassign": "off",
+      "eol-last": "off",
+      "arrow-parens": "off"
     }
-}
+  }
