+++ package.json
@@ -1,13 +1,26 @@
 {
   "name": "tic-tac-toe-js",
   "version": "1.0.0",
-  "main": "script.js",
+  "main": "'./src/index.js'",
+  "scripts": {
+    "build": "webpack --mode development",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "watch": "webpack --progress --watch"
+  },
   "repository": "https://github.com/geraldgsh/tic-tac-toe-js/",
-  "author": "Izaias Neto <izaiasneto4@gmail.com>",
+  "author": "Gerald Goh",
   "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/geraldgsh/tic-tac-toe-js/issues"
+  },
+  "homepage": "https://github.com/geraldgsh/tic-tac-toe-js#readme",
   "devDependencies": {
+    "@babel/preset-env": "^7.8.6",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.20.0"
+    "eslint-plugin-import": "^2.20.0",
+    "jest": "^25.1.0",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11"
   }
 }
