+++ src/index.js
@@ -1,0 +1,3 @@
+import * as gameEngine from './modules/gameEngine';
+
+gameEngine;
