+++ src/modules/board.js
@@ -1,0 +1,19 @@
+const playBoard = (() => {
+  const gridBoard = ['', '', '', '', '', '', '', '', ''];
+  const winningCombination = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 4, 8],
+    [2, 4, 6],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+  ];
+  return {
+    gridBoard,
+    winningCombination,
+  };
+})();
+
+module.exports = { playBoard, default: playBoard };
