+++ src/modules/playerGenerator.js
@@ -1,0 +1,8 @@
+const playerGenerator = (name, mark, wins, playArr) => ({
+  name,
+  mark,
+  wins,
+  playArr,
+});
+
+module.exports = { playerGenerator, default: playerGenerator };
